--HG--
branch : popiii_colorfield

+++ src/enzo/GetUnits.C
@@ -29,7 +29,7 @@ int CosmologyGetUnits(float *DensityUnits, float *LengthUnits,
 
 int GetUnits(float *DensityUnits, float *LengthUnits,
 	     float *TemperatureUnits, float *TimeUnits,
-	     float *VelocityUnits, float *MassUnits, FLOAT Time)
+	     float *VelocityUnits, double *MassUnits, FLOAT Time)
 
 {
   /* If using cosmology, get cosmology units */
